Split MainPage.cs
